-Checkpoint commit- Additions made: The Models were moved to their own class library, I just wanted to demonstrate that I know how to. UserAccountVM was created to help shape the data coming in from the api. AccountUserController was created to help funnel the api into the razor page(s).
