+++ AccountsTheGathering.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 16.0.31702.278
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AccountsTheGathering", "AccountsTheGathering\AccountsTheGathering.csproj", "{681F7A59-1D46-4CAD-B9AB-4B85CC820010}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AccountsTheGathering_Models", "AccountsTheGathering_Models\AccountsTheGathering_Models.csproj", "{61625D4F-3444-4B98-A892-C059E53D72DD}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{681F7A59-1D46-4CAD-B9AB-4B85CC820010}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{681F7A59-1D46-4CAD-B9AB-4B85CC820010}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{681F7A59-1D46-4CAD-B9AB-4B85CC820010}.Release|Any CPU.Build.0 = Release|Any CPU
+		{61625D4F-3444-4B98-A892-C059E53D72DD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{61625D4F-3444-4B98-A892-C059E53D72DD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{61625D4F-3444-4B98-A892-C059E53D72DD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{61625D4F-3444-4B98-A892-C059E53D72DD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
