+++ AccountsTheGathering/AccountsTheGathering.csproj
@@ -17,4 +17,8 @@
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Razor.RuntimeCompilation" Version="5.0.10" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\AccountsTheGathering_Models\AccountsTheGathering_Models.csproj" />
+  </ItemGroup>
+
 </Project>
