+++ AccountsTheGathering/Controllers/HomeController.cs
@@ -1,17 +1,27 @@
-﻿using AccountsTheGathering.Models;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
+using AccountsTheGathering_Models.ViewModels;
 
 namespace AccountsTheGathering.Controllers
 {
     public class HomeController : Controller
     {
-        public HomeController() { }
+        /// <summary>
+        /// CTORs are generated by default, but it's okay to have here just in case.
+        /// </summary>
+        public HomeController()
+        {
+            
+        }
 
+        /// <summary>
+        /// The main goods of this page.
+        /// </summary>
+        /// <returns></returns>
         public IActionResult Index()
         {
             return View();
