+++ AccountsTheGathering/Views/Shared/Error.cshtml
@@ -1,4 +1,4 @@
-﻿@model ErrorViewModel
+﻿@model AccountsTheGathering_Models.ViewModels.ErrorViewModel
 @{
     ViewData["Title"] = "Error";
 }
