+++ AccountsTheGathering/Views/_ViewImports.cshtml
@@ -1,3 +1,3 @@
 ﻿@using AccountsTheGathering
-@using AccountsTheGathering.Models
+@using AccountsTheGathering_Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
